Rename file and directory
