+++ .travis.yml
@@ -3,7 +3,7 @@ language: python
 services:
   - docker
 before_install:
-  - docker build -t roboticbase/fiware_ros_turtlebot3_bridge .
+  - docker build -t roboticbase/fiware_ros_bridge .
 script:
-  - docker run roboticbase/fiware_ros_turtlebot3_bridge /bin/bash -c "pip install -r src/fiware_ros_turtlebot3_bridge/requirements/develop.txt;source /opt/ros/kinetic/setup.bash;catkin_make;source devel/setup.bash;flake8 src/fiware_ros_turtlebot3_bridge"
-  - docker run roboticbase/fiware_ros_turtlebot3_bridge /bin/bash -c "pip install -r src/fiware_ros_turtlebot3_bridge/requirements/develop.txt;source /opt/ros/kinetic/setup.bash;catkin_make;source devel/setup.bash;catkin_make run_tests;catkin_test_results"
+  - docker run roboticbase/fiware_ros_bridge /bin/bash -c "pip install -r src/fiware_ros_bridge/requirements/develop.txt;source /opt/ros/kinetic/setup.bash;catkin_make;source devel/setup.bash;flake8 src/fiware_ros_bridge"
+  - docker run roboticbase/fiware_ros_bridge /bin/bash -c "pip install -r src/fiware_ros_bridge/requirements/develop.txt;source /opt/ros/kinetic/setup.bash;catkin_make;source devel/setup.bash;catkin_make run_tests;catkin_test_results"
