--- launch/fiware_ros_turtlebot3_bridge.launch
@@ -1,14 +1,0 @@
-<launch>
-  <node name="turtlebot3_cmd" pkg="fiware_ros_turtlebot3_bridge" type="turtlebot3_cmd.py" output="screen">
-    <rosparam command="delete" param=""/>
-    <rosparam command="load" file="$(find fiware_ros_turtlebot3_bridge)/config/mqtt.yaml"/>
-    <rosparam command="load" file="$(find fiware_ros_turtlebot3_bridge)/config/turtlebot3_cmd.yaml"/>
-    <param name="mqtt/cafile" value="$(find fiware_ros_turtlebot3_bridge)/certs/DST_Root_CA_X3.pem"/>
-  </node>
-  <node name="turtlebot3_attrs" pkg="fiware_ros_turtlebot3_bridge" type="turtlebot3_attrs.py" output="screen">
-    <rosparam command="delete" param=""/>
-    <rosparam command="load" file="$(find fiware_ros_turtlebot3_bridge)/config/mqtt.yaml"/>
-    <rosparam command="load" file="$(find fiware_ros_turtlebot3_bridge)/config/turtlebot3_attrs.yaml"/>
-    <param name="mqtt/cafile" value="$(find fiware_ros_turtlebot3_bridge)/certs/DST_Root_CA_X3.pem"/>
-  </node>
-</launch>
